tests: Move python sign+verify test into api_test.py
